requirements.txt
requests~=2.31.0
termcolor~=2.3.0
web.py~=0.62
fake-useragent~=1.3.0
openai~=1.61.1
python-dotenv~=1.0.1